Module lock_keeper_tests::test_suites::config_files

Config file tests that depend on files to exist at specific paths. If the paths listed in these tests change, the tests will fail.

Constants

CLIENT_CONFIG_NO_KEY 🔒

CLIENT_CONFIG_WITH_KEY 🔒

SAMPLE_PRIVATE_KEY 🔒

SAMPLE_SSE_KEY 🔒

SERVER_CONFIG_NO_KEY 🔒

SERVER_CONFIG_NO_SSE_KEY 🔒

SERVER_CONFIG_WITH_KEY 🔒

Functions

client_config_with_file_private_key_works 🔒

client_config_with_manual_private_key_works 🔒

client_config_without_private_key_fails 🔒

run_tests

server_config_with_file_private_key_works 🔒

server_config_with_manual_private_key_works 🔒

server_config_with_manual_remote_storage_key_works 🔒

server_config_without_private_key_fails 🔒

server_config_without_remote_storage_key_fails 🔒