Module lock_keeper_tests::test_suites::database::audit_event

Integration tests for audit event objects in the database

Constants

HOW_MANY_SECRETS ğŸ”’

NUM_LOGS ğŸ”’

NUM_SAMPLE ğŸ”’

Functions

after_date_filter_works ğŸ”’

before_date_filter_works ğŸ”’

check_after_date ğŸ”’

check_before_date ğŸ”’

compare_actions ğŸ”’

compare_key_ids ğŸ”’

compare_request_ids ğŸ”’

We expect all events to have the request_id we specified. Check this.

create_random_audit_event ğŸ”’

Create a single random audit event and store it in our DB. This function samples from the passed vectors of client actions and key IDs.

create_random_audit_events ğŸ”’

Create NUM_LOGS random audit events and store them in our database. Return the KeyIds and Uuids (request IDs) assigned to these audit events.

event_type_filter_works ğŸ”’

Check that the DB test filters work by ensuring queries to the DB with specific EventTypes return only the specified types of audit event log events.

key_id_filter_works ğŸ”’

request_id_filter_works ğŸ”’

run_tests

store_audit_event_identity ğŸ”’

Tests that storing an event returns the same event back out.