Struct lock_keeper_tests::test_suites::database::TestDatabase

pub struct TestDatabase {
    pub db: PostgresDB,
}

Fields
db: PostgresDB
Implementations

impl TestDatabase

fn blob_test_data() -> Vec<u8>

pub async fn connect() -> Result<Self>

fn create_test_master_key(rng: &mut StdRng) -> Result<MasterKey>

Create a master key for testing using random bytes.

pub fn create_test_storage_key(
&self,
user_id: &UserId
) -> Result<(Encrypted<StorageKey>, MasterKey)>

Create a storage key for a test user.

pub async fn create_test_user(&self) -> Result<Account>

Creates a test user with a randomized name and returns the id and account name.

async fn create_random_arbitrary_secrets(
&self,
n: usize,
account: &Account
) -> Result<Vec<KeyId>>

Generate and store n random secrets in our database.

async fn store_server_encrypted_blob(
&self,
rng: &mut StdRng,
account: &Account
) -> Result<(KeyId, RemoteStorageKey)>

Store a data blog in database. These are the steps server takes to encrypt blob.

async fn add_arbitrary_secret(
&self,
rng: &mut StdRng,
storage_key: &StorageKey,
account: &Account
) -> Result<KeyId>

Store a new arbitrary secret in database.

async fn import_signing_key(
&self,
rng: &mut StdRng,
account: &Account
) -> Result<KeyId>

async fn remote_generate_signing_key(
&self,
rng: &mut StdRng,
account: &Account
) -> Result<KeyId>