Module lock_keeper_tests::test_suites::end_to_end::operations

Functions

authenticate ğŸ”’

Authenticate to the LockKeeper key server and get a LockKeeperClient.

check_audit_events ğŸ”’

Fetch relevant events based on request_id. Ensure that some event on the list matches the expected_status and expected_action. Note: we cannot rely on the ordering of the audit events.

compare_errors ğŸ”’

Helper to compare result errors.

compare_status_errors ğŸ”’

Helper to compare status errors.

generate_fake_key_id ğŸ”’

Generate a fake key ID to test retrieve/export failure cases.

import_signing_key ğŸ”’

Authenticate and import a signing key.