Module lock_keeper_tests::utils

Various testing utilities

Enums

TestResult

Constants

RNG_SEED

Functions

random_bytes

Generate a sequence of random bytes with the given length.

report_test_results

run_test_case

Runs a test case and manually handles any panics triggered by assert macros.

server_registration

Locally simulates OPAQUE registration to get a valid ServerRegistration for remaining tests.

tagged

Add random text to the end of a string

wait_for_server