Crate lock_keeper_tests

Modules

config

Test config types

error

test_suites

utils

Various testing utilities

Macros

run_parallel

Pass a list of async test functions with parameters included. All test functions must return [Result<()>]. Tests will run in parallel and report the names of any failing test.

Structs

Cli

Enums

TestType

Functions

main

run ğŸ”’

run_integration_tests ğŸ”’

run_tests ğŸ”’