Module lock_keeper_tests::test_suites::end_to_end::operations

Functions

authenticate ğŸ”’

Authenticate to the LockKeeper key server and get a LockKeeperClient.

check_audit_events ğŸ”’

Make sure the last audit log for the tested operation has the correct status and action.

compare_errors ğŸ”’

Helper to compare result errors.

export ğŸ”’

Authenticate and export an arbitrary secret by KeyId.

export_signing_key ğŸ”’

Authenticate and export a signing key by KeyId.

generate ğŸ”’

Authenticate and generate an arbitrary key.

generate_fake_key_id ğŸ”’

Generate a fake key ID to test retrieve/export failure cases.

import_signing_key ğŸ”’

Authenticate and import a signing key.

register ğŸ”’

Register to the LockKeeper key server.

remote_generate ğŸ”’

Authenticate and generate a signing key server-side

remote_sign_bytes ğŸ”’

Authenticate and sign bytes with a remote generated signing key

retrieve ğŸ”’

Authenticate and retrieve an arbitrary key by KeyId.