Module lock_keeper_tests::test_suites::session_cache

Session cache integration tests.

Structs

TestState ğŸ”’

Constants

FILLER ğŸ”’

SEED ğŸ”’

Functions

get_key_back ğŸ”’

Ensure we can get the key back without encountering expiration error.

get_temp_account_id ğŸ”’

get_temp_remote_key ğŸ”’

get_temp_session_id ğŸ”’

get_temp_session_key ğŸ”’

key_does_not_exist ğŸ”’

key_expired ğŸ”’

Test key expiration logic when enough time has passed that the key should be expired.

key_expired2 ğŸ”’

Key expires after a short time.

new_cache

Create a new test database with random characters appended to name. This is what you probably want to use.

overwrite_existing_key ğŸ”’

Insert key twice and ensure code runs all the way.

run_tests

test_state ğŸ”’